data/snippets/requirements.txt
python-dotenv==1.1.0

requests==2.32.3

boto3==1.37.33
requests==2.32.3

pymongo==4.12.0
mongoengine==0.29.1

lumaai==1.7.3
mistralai==1.6.0
